fix(experimental): Move experimental patterns into the correct directory
This follows #1418 where we want to be more explicit about what parts of the library may not have a stable API.
