+++ src/patterns/index.ts
@@ -1,7 +1,5 @@
 export * from "./api-lambda";
 export * from "./api-multiple-lambdas";
 export * from "./scheduled-lambda";
-export * from "./sns-lambda";
-export * from "./kinesis-lambda";
 export * from "./ec2-app";
 export * from "./scheduled-ecs-task";
